move scenes into scenes folder
